Add files via upload
For anyone wondering how I populated the mySQL database with the GTFS data. I found this piece of code that did the job. I kept the original writers README file; if anyone wants to read it. Some of the original code was written for python 2.7ish so it was a bit outdated and it was missing some tables. Important note: The database will have to be created in mySQL workbench before the python code should be used.  The code only populates the database and converts arrival times and departure times a usable format.